Define MockFileSystemService interface for mock Filesystem activities consumed by usecases unit tests.
