+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/usecases/AnalyzeCopybookCaching.java
@@ -1,3 +1,18 @@
+/*
+ *
+ *  Copyright (c) 2020 Broadcom.
+ *  The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
+ *
+ *  This program and the accompanying materials are made
+ *  available under the terms of the Eclipse Public License 2.0
+ *  which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ *  SPDX-License-Identifier: EPL-2.0
+ *
+ *  Contributors:
+ *    Broadcom, Inc. - initial API and implementation
+ *
+ */
 package com.ca.lsp.cobol.usecases;
 
 import com.broadcom.lsp.cdi.LangServerCtx;
@@ -7,7 +22,8 @@ import com.broadcom.lsp.domain.cobol.databus.model.CopybookStorable;
 import com.broadcom.lsp.domain.common.model.Position;
 import com.ca.lsp.cobol.ConfigurableTest;
 import com.ca.lsp.cobol.positive.CobolText;
-import com.ca.lsp.cobol.service.mocks.MockWorkspaceService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemServiceImpl;
 import com.ca.lsp.core.cobol.model.CopybookDefinition;
 import com.ca.lsp.core.cobol.preprocessor.sub.copybook.AnalyseCopybookTask;
 import com.ca.lsp.core.cobol.preprocessor.sub.util.impl.MultiMapSerializableHelper;
@@ -42,6 +58,9 @@ public class AnalyzeCopybookCaching extends ConfigurableTest {
   private final DefaultDataBusBroker databus =
       LangServerCtx.getInjector().getInstance(DefaultDataBusBroker.class);
 
+  private final MockFileSystemService mockFileSystemService =
+      LangServerCtx.getInjector().getInstance(MockFileSystemServiceImpl.class);
+
   @Before
   public void init() {
     initParagraphDefinitions();
@@ -65,8 +84,7 @@ public class AnalyzeCopybookCaching extends ConfigurableTest {
   }
 
   private void initWorkspaceService() {
-    MockWorkspaceService mockWorkspaceService = new MockWorkspaceService(databus);
-    mockWorkspaceService.setCopybooks(
+    mockFileSystemService.setCopybooks(
         () -> Collections.singletonList(new CobolText(COPYBOOK_NAME, COPYBOOK_CONTENT)));
   }
