+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/usecases/TestCopybookWithRecursiveDependencyIsDetected.java
@@ -16,7 +16,8 @@ package com.ca.lsp.cobol.usecases;
 
 import com.broadcom.lsp.cdi.LangServerCtx;
 import com.ca.lsp.cobol.positive.CobolText;
-import com.ca.lsp.cobol.service.mocks.MockWorkspaceService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemServiceImpl;
 import org.eclipse.lsp4j.Range;
 import org.junit.Test;
 
@@ -43,9 +44,9 @@ public class TestCopybookWithRecursiveDependencyIsDetected extends NegativeUseCa
   public TestCopybookWithRecursiveDependencyIsDetected() {
     super(TEXT);
 
-    MockWorkspaceService workspaceService =
-        LangServerCtx.getInjector().getInstance(MockWorkspaceService.class);
-    workspaceService.setCopybooks(
+    MockFileSystemService mockFileSystemService =
+        LangServerCtx.getInjector().getInstance(MockFileSystemServiceImpl.class);
+    mockFileSystemService.setCopybooks(
         () -> Collections.singletonList(new CobolText("RECURSIVE-COPY", RECURSIVE_COPY)));
   }
