+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/usecases/TestVariableStructureInCopybooksWithMissingLevels.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Broadcom.
+ * Copyright (c) 2020 Broadcom.
  * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
  * This program and the accompanying materials are made
@@ -16,7 +16,8 @@ package com.ca.lsp.cobol.usecases;
 
 import com.broadcom.lsp.cdi.LangServerCtx;
 import com.ca.lsp.cobol.positive.CobolText;
-import com.ca.lsp.cobol.service.mocks.MockWorkspaceService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemServiceImpl;
 import org.junit.Test;
 
 import java.util.Collections;
@@ -46,9 +47,10 @@ public class TestVariableStructureInCopybooksWithMissingLevels extends PositiveU
   public TestVariableStructureInCopybooksWithMissingLevels() {
     super(TEXT);
 
-    MockWorkspaceService workspaceService =
-        LangServerCtx.getInjector().getInstance(MockWorkspaceService.class);
-    workspaceService.setCopybooks(() -> Collections.singletonList(new CobolText("STRUCT", STRUCT)));
+    MockFileSystemService mockFileSystemService =
+        LangServerCtx.getInjector().getInstance(MockFileSystemServiceImpl.class);
+    mockFileSystemService.setCopybooks(
+        () -> Collections.singletonList(new CobolText("STRUCT", STRUCT)));
   }
 
   @Override
