+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/usecases/TestVariableStructureIsBuiltWithCopybooks.java
@@ -16,7 +16,8 @@ package com.ca.lsp.cobol.usecases;
 
 import com.broadcom.lsp.cdi.LangServerCtx;
 import com.ca.lsp.cobol.positive.CobolText;
-import com.ca.lsp.cobol.service.mocks.MockWorkspaceService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemService;
+import com.ca.lsp.cobol.service.mocks.MockFileSystemServiceImpl;
 import org.junit.Test;
 
 import java.util.Collections;
@@ -46,9 +47,9 @@ public class TestVariableStructureIsBuiltWithCopybooks extends PositiveUseCase {
 
   public TestVariableStructureIsBuiltWithCopybooks() {
     super(TEXT);
-    MockWorkspaceService workspaceService =
-        LangServerCtx.getInjector().getInstance(MockWorkspaceService.class);
-    workspaceService.setCopybooks(
+    MockFileSystemService mockFileSystemService =
+        LangServerCtx.getInjector().getInstance(MockFileSystemServiceImpl.class);
+    mockFileSystemService.setCopybooks(
         () -> Collections.singletonList(new CobolText("COPYBOOK-CONTENT", COPYBOOK_CONTENT)));
   }
